IA-Immo
Our project consists in determining for which surface in m² we can claim a monthly rent that would suit a client wishing to rent an apartment or a house in the Hauts-de-Seine, the values taken for this project are from July 2021.

As our goal is to return a value, we face a regression problem in supervised learning because our algorithm trains on labeled data, it knows that the table x corresponds to monthly rents and that the table y corresponds to surfaces in m².

The code starts by importing the three libraries necessary to achieve our goal: Scikit-Learn (sklearn), NumPy and Matplotlib

Since the Scikit-Learn library contains many models, we specify that we want to import a model that allows us to perform linear regressions with the instruction from sklearn.linear_model import LinearRegression.

When importing NumPy, we use the keyword as to indicate that we will call it through the alias np, which is shorter to write than numpy. We do the same with the mathplotlib.pyplot library by renaming it plt, which is much shorter to write.

The next step is to create two arrays of data used by the algorithm. We use for that the array() method of NumPy. This method takes as a parameter a list of numerical values which are enclosed in square brackets, each value being separated from another by a comma. The two arrays contain five values each. We put each array in two variables: x and y. The table x contains the values of the rents in € and the values of y contain the values of the surfaces in square meters.

Then, we create the model, here the LinearRegression() function. We put this model in a variable that we call model. The training of the model, characterized by the calculation of the directing coefficient and the y-intercept, is done by using the fit() method on the x and y arrays passed as parameters. It is necessary to modify the dimension of the x array. We put it in two dimensions by using the function reshape(-1, 1) as follows: x.reshape(-1, 1).

We then apply the score() method to the model data to determine the prediction score of the model, i.e. at what point the line passes closest to the points. The first print indicates a score of 0.99, which means 99% success rate. The error is therefore very small, the model is performing well.

The method predict() allows to find the value y for a given value x. It takes as parameter a rent value and returns a value of apartment size in square meters. In our case, it tells us that for 1000 € of rent, we can expect an apartment of 35.39 m².

Finally, we move to the visualization. We use the scatter() method of matplolib.pyplot. Each point is placed according to the coordinates found in the x and y arrays. We use the plot() function to draw a line. To do this, we must indicate on the x axis the values of the rents. On the y-axis, we indicate the values of the apartment sizes. For this reason we write model.predict(x.reshape(-1, 1)) as the second parameter. As a third parameter, we indicate the color of the line, here in red.

Don't forget to give a title to your graph with the title("") method and don't forget the title of the axes with the xlabel("") method for the x-axis and ylabel("") for the y-axis.
And finally, we use the show() function to display the graph.